REL-690562 do not remove workspace after fail

+++ Source/kCura.IntegrationPoints.FunctionalTests/Functional/CI/TestsBase.cs
@@ -1,4 +1,5 @@
 ﻿using System.Net.Http;
+using NUnit.Framework;
 using Polly;
 using Polly.Retry;
 using Relativity.IntegrationPoints.Tests.Functional.Helpers;
@@ -49,7 +50,7 @@ namespace Relativity.IntegrationPoints.Tests.Functional.CI
         protected override void OnTearDownFixture()
         {
             base.OnTearDownFixture();
-            if (_existingWorkspaceArtifactID == 0)
+            if (_existingWorkspaceArtifactID == 0 && TestContext.CurrentContext.Result.FailCount == 0)
             {
                 RelativityFacade.Instance.DeleteWorkspace(Workspace);
             }
